[FEAT/#78] - DEV pull 받고 DTO 통일
