+++ app/src/main/java/com/example/assu_fe_app/data/dto/chatting/request/CreateChatRoomRequestDto.kt
@@ -4,6 +4,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class CreateChatRoomRequestDto(
-    val storeId: Long,
+    val adminId: Long,
     val partnerId: Long
 )
