+++ app/src/main/java/com/example/assu_fe_app/data/dto/location/LocationAdminPartnerSearchResultItem.kt
@@ -7,6 +7,5 @@ data class LocationAdminPartnerSearchResultItem(
     val isPartnered: Boolean,
     val paperId: Long?,
     val term: String?,
-    val storeId: Long? = null,
     val partnershipId: Long? = null
 )
