+++ app/src/main/java/com/example/assu_fe_app/data/dto/partnership/response/GetProposalPartnerListResponseDto.kt
@@ -1,9 +1,6 @@
 package com.example.assu_fe_app.data.dto.partnership.response
 
 import com.example.assu_fe_app.domain.model.admin.GetProposalPartnerListModel
-import com.example.assu_fe_app.domain.model.admin.PartnershipGoodsModel
-import com.example.assu_fe_app.domain.model.admin.PartnershipOptionModel
-import com.example.assu_fe_app.domain.model.admin.ProposalPartnerDetailsModel
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
