+++ app/src/main/java/com/example/assu_fe_app/data/dto/partnership/response/GoodsResponseDto.kt
@@ -1,6 +1,6 @@
 package com.example.assu_fe_app.data.dto.partnership.response
 
-import com.example.assu_fe_app.domain.model.partnership.GoodsModel
+import com.example.assu_fe_app.domain.model.partnership.PartnershipGoodsModel
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
@@ -10,10 +10,8 @@ data class GoodsResponseDto(
     val price: Int?,            // 없으면 제거
     val description: String?    // 없으면 제거
 ) {
-    fun toModel() = GoodsModel(
-        id = id ?: -1L,
-        name = name.orEmpty(),
-        price = price ?: 0,
-        description = description.orEmpty()
+    fun toModel() = PartnershipGoodsModel(
+        goodsId = id ?: -1L,
+        goodsName = name.orEmpty(),
     )
 }
