+++ app/src/main/java/com/example/assu_fe_app/data/repositoryImpl/partnership/PartnershipRepositoryImpl.kt
@@ -6,7 +6,7 @@ import com.example.assu_fe_app.data.service.notification.NotificationService
 import com.example.assu_fe_app.data.service.partnership.PartnershipService
 import com.example.assu_fe_app.domain.model.admin.GetProposalAdminListModel
 import com.example.assu_fe_app.domain.model.admin.GetProposalPartnerListModel
-import com.example.assu_fe_app.domain.model.partnership.PartnershipDetailModel
+import com.example.assu_fe_app.domain.model.partnership.ProposalPartnerDetailsModel
 import com.example.assu_fe_app.util.RetrofitResult
 import com.example.assu_fe_app.util.apiHandler
 import jakarta.inject.Inject
@@ -27,7 +27,7 @@ class PartnershipRepositoryImpl @Inject constructor(
         )
 
     override suspend fun getPartnership(partnershipId: Long)
-            : RetrofitResult<PartnershipDetailModel> =
+            : RetrofitResult<ProposalPartnerDetailsModel> =
         apiHandler(
             execute = { api.getPartnership(partnershipId) },
             mapper = { dto -> dto.toModel() }
