+++ app/src/main/java/com/example/assu_fe_app/data/service/location/LocationService.kt
@@ -48,4 +48,20 @@ interface LocationService {
         @Query("lng4") lng4: Double,
         @Query("lat4") lat4: Double
     ): BaseResponse<List<AdminMapResponseDto>>
+
+    @GET("/map/search")
+    suspend fun searchStores(
+        @Query("searchKeyword") keyword: String
+    ): BaseResponse<List<StoreMapResponseDto>>
+
+    @GET("/map/search") // admin 입장에서 검색한 결과
+    suspend fun searchPartners(
+        @Query("searchKeyword") keyword: String
+    ): BaseResponse<List<PartnerMapResponseDto>>
+
+    // 파트너용 검색 API
+    @GET("/map/search") // partner 입장에서 검색한 결과
+    suspend fun searchAdmins(
+        @Query("searchKeyword") keyword: String
+    ): BaseResponse<List<AdminMapResponseDto>>
 }
