--- app/src/main/java/com/example/assu_fe_app/domain/model/admin/PartnershipGoodsModel.kt
@@ -1,6 +1,0 @@
-package com.example.assu_fe_app.domain.model.admin
-
-data class PartnershipGoodsModel(
-    val goodsId: Long,
-    val goodsName: String,
-)
