+++ app/src/main/java/com/example/assu_fe_app/domain/model/location/StudentMapSearchResultModel.kt
@@ -1,0 +1,4 @@
+package com.example.assu_fe_app.domain.model.location
+
+class StudentMapSearchResultModel {
+}
