--- app/src/main/java/com/example/assu_fe_app/domain/model/map/StudentMapSearchResultModel.kt
@@ -1,4 +1,0 @@
-package com.example.assu_fe_app.domain.model.map
-
-class StudentMapSearchResultModel {
-}
