--- app/src/main/java/com/example/assu_fe_app/domain/model/partnership/GoodsModel.kt
@@ -1,8 +1,0 @@
-package com.example.assu_fe_app.domain.model.partnership
-
-data class GoodsModel(
-    val id: Long,
-    val name: String,
-    val price: Int,
-    val description: String
-)
