+++ app/src/main/java/com/example/assu_fe_app/domain/model/partnership/PartnershipGoodsModel.kt
@@ -1,0 +1,6 @@
+package com.example.assu_fe_app.domain.model.partnership
+
+data class PartnershipGoodsModel(
+    val goodsId: Long,
+    val goodsName: String
+)
