+++ app/src/main/java/com/example/assu_fe_app/domain/model/partnership/PartnershipOptionModel.kt
@@ -1,11 +1,14 @@
 package com.example.assu_fe_app.domain.model.partnership
 
+import com.example.assu_fe_app.data.dto.partnership.response.CriterionType
+import com.example.assu_fe_app.data.dto.partnership.response.OptionType
+
 data class PartnershipOptionModel(
-    val optionType: String,
-    val criterionType: String,
+    val optionType: OptionType,
+    val criterionType: CriterionType,
     val people: Int,
-    val cost: Int,
+    val cost: Long,
     val category: String,
-    val discountRate: Int,
-    val goods: List<GoodsModel>
+    val discountRate: Long,
+    val goods: List<PartnershipGoodsModel>
 )
