+++ app/src/main/java/com/example/assu_fe_app/domain/usecase/location/UserSearchStoreByKeywordUseCase.kt
@@ -1,16 +1,16 @@
-package com.example.assu_fe_app.domain.usecase.map
+package com.example.assu_fe_app.domain.usecase.location
 
 import com.example.assu_fe_app.data.dto.location.LocationUserSearchResultItem
-import com.example.assu_fe_app.data.repository.map.MapRepository
+import com.example.assu_fe_app.data.repository.location.LocationRepository
 import com.example.assu_fe_app.util.RetrofitResult
 import javax.inject.Inject
 
 class UserSearchStoreByKeywordUseCase @Inject constructor(
-    private val mapRepository: MapRepository
+    private val locationRepository: LocationRepository
 ) {
     suspend operator fun invoke(keyword: String)
     : RetrofitResult<List<LocationUserSearchResultItem>> {
-        return mapRepository.searchStores(keyword)
+        return locationRepository.searchStores(keyword)
     }
 
 }
