+++ app/src/main/java/com/example/assu_fe_app/domain/usecase/partnership/GetPartnershipUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.assu_fe_app.domain.usecase.partnership
 
 import com.example.assu_fe_app.data.repository.partnership.PartnershipRepository
-import com.example.assu_fe_app.domain.model.partnership.PartnershipDetailModel
+import com.example.assu_fe_app.domain.model.partnership.ProposalPartnerDetailsModel
 import com.example.assu_fe_app.util.RetrofitResult
 import jakarta.inject.Inject
 
@@ -9,6 +9,6 @@ class GetPartnershipUseCase @Inject constructor(
     private val repo: PartnershipRepository
 ) {
     suspend operator fun invoke(partnershipId: Long)
-            : RetrofitResult<PartnershipDetailModel> =
+            : RetrofitResult<ProposalPartnerDetailsModel> =
         repo.getPartnership(partnershipId)
 }
