+++ app/src/main/java/com/example/assu_fe_app/presentation/common/contract/adapter/PartnershipContractAdapter.kt
@@ -6,7 +6,6 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.assu_fe_app.R
 import com.example.assu_fe_app.data.dto.partner_admin.home.PartnershipContractItem
 import com.example.assu_fe_app.databinding.ItemPartnershipContentListBinding
-import com.example.assu_fe_app.domain.model.partnership.PartnershipOptionModel
 
 
 class PartnershipContractAdapter (
