+++ app/src/main/java/com/example/assu_fe_app/presentation/common/location/adapter/AdminParterLocationAdapter.kt
@@ -19,7 +19,7 @@ class AdminPartnerLocationAdapter :
     ) : RecyclerView.ViewHolder(binding.root) {
 
         fun bind(item: LocationAdminPartnerSearchResultItem, isLastItem: Boolean) {
-            binding.tvItemAdminPartnerLocationSearchResultItemShopName.text = item.name
+            binding.tvItemAdminPartnerLocationSearchResultItemShopName.text = item.shopName
 
             if (item.isPartnered) {
                 binding.tvItemAdminPartnerLocationSearchResultItemPartnered.visibility = View.VISIBLE
