+++ app/src/main/java/com/example/assu_fe_app/presentation/partner/home/PartnerHomeFragment.kt
@@ -216,7 +216,7 @@ class PartnerHomeFragment :
         binding.viewPartnerHomeCardBg.setOnClickListener {
             val req = CreateChatRoomRequestDto(
                 //TODO : 유저 정보 받아오기
-                storeId = 1L,
+                adminId = 1L,
                 partnerId = 5L
             )
             chattingViewModel.createRoom(req)
