+++ app/src/main/java/com/example/assu_fe_app/ui/map/AdminPartnerKeyWordSearchViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.assu_fe_app.data.dto.location.LocationAdminPartnerSearchResultItem
-import com.example.assu_fe_app.domain.usecase.map.AdminSearchPartnerByKeywordUseCase
-import com.example.assu_fe_app.domain.usecase.map.PartnerSearchAdminByKeywordUseCase
+import com.example.assu_fe_app.domain.usecase.location.AdminSearchPartnerByKeywordUseCase
+import com.example.assu_fe_app.domain.usecase.location.PartnerSearchAdminByKeywordUseCase
 import com.example.assu_fe_app.util.RetrofitResult
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -33,17 +33,12 @@ constructor(
         // 여기서 로그인된 사용자의 유저 정보를 불러서 업데이트
     }
 
-
     fun searchPartners(keyword: String){
-        viewModelScope.launch{
+        viewModelScope.launch {
             when (val result = adminSearchPartnerCase(keyword)) {
-                is RetrofitResult.Success -> {
+                is RetrofitResult.Success<List<LocationAdminPartnerSearchResultItem>> -> {
                     _contentList.value = result.data
-                    if(result.data.isEmpty()){
-                        _isEmptyList.value = true
-                    }else{
-                        _isEmptyList.value = false
-                    }
+                    _isEmptyList.value = result.data.isEmpty()
                 }
                 is RetrofitResult.Error -> {
                     Log.d("❌", "Error : ${result.exception.message}")
@@ -51,7 +46,6 @@ constructor(
                 is RetrofitResult.Fail -> {
                     Log.d("❌", "Fail : ${result.message}")
                 }
-
             }
         }
     }
@@ -59,23 +53,16 @@ constructor(
     fun searchAdmins(keyword: String){
         viewModelScope.launch {
             Log.d("AdminPartnerKeyWordSearchViewModel", "Partner 기준 Admin 찾기 함수 호출 중")
-            when( val result = partnerSearchAdminCase(keyword)){
-
-                is RetrofitResult.Success -> {
+            when (val result = partnerSearchAdminCase(keyword)) {
+                is RetrofitResult.Success<List<LocationAdminPartnerSearchResultItem>> -> {
                     _contentList.value = result.data
-                    if(result.data.isEmpty()){
-                        _isEmptyList.value = true
-                    }else{
-                        _isEmptyList.value = false
-                    }
+                    _isEmptyList.value = result.data.isEmpty()
                 }
                 is RetrofitResult.Error -> {
                     Log.d("❌", "Error : ${result.exception.message}")
                 }
                 is RetrofitResult.Fail -> {
                     Log.d("❌", "Fail : ${result.message}")
-
-
                 }
             }
         }
