+++ app/src/main/java/com/example/assu_fe_app/ui/map/UserLocationSearchViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.assu_fe_app.data.dto.location.LocationUserSearchResultItem
-import com.example.assu_fe_app.domain.usecase.map.UserSearchStoreByKeywordUseCase
+import com.example.assu_fe_app.domain.usecase.location.UserSearchStoreByKeywordUseCase
 import com.example.assu_fe_app.util.RetrofitResult
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
